+++ utils/utils.h
@@ -6,7 +6,7 @@
 /*   By: asenat <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/09/24 11:30:35 by asenat            #+#    #+#             */
-/*   Updated: 2018/09/24 13:31:58 by asenat           ###   ########.fr       */
+/*   Updated: 2018/09/24 18:32:22 by asenat           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,6 +14,7 @@
 # define UTILS_H
 
 # include <sys/stat.h>
+# include <stdint.h>
 
 typedef struct 	s_file
 {
@@ -22,7 +23,7 @@ typedef struct 	s_file
 	struct stat	*stats;
 }				t_file;
 
-int		open_file(const char* filename, int oflags, t_file *file);
-void	close_file(t_file* file);
+uint8_t	open_file(const char *filename, int oflags, t_file *file);
+void	close_file(t_file *file);
 
 #endif
